separate object_rt and object_ge
