+++ taichi_three/common.py
@@ -1,2 +1,5 @@
 EPS = 1e-3
 INF = 1e5
+
+class TaichiClass:
+    is_taichi_class = True
